Worked on v1.0.
This commit finishes all functionaly of the bot.

All the functions have been rewritten using promises in order to improve the stability of the code.
When something goes wrong, an error will be printed to the console and the execution will retry on the next day.

This commit also adds a Continous Integration (powered by Jest) to test code before deploying it using Heroku.